More refactoring fixes
