+++ src/app/components/windows/WindowRatings.vue
@@ -73,8 +73,8 @@ import { onBeforeUnmount, onMounted, reactive, ref, watch } from 'vue'
 import { api } from '@app/api/api'
 import { useWindowsStore } from '@app/stores/windowsStore'
 import WinWindow from '@app/components/basic/WinWindow.vue'
-import type WinList from '@app/components/WinList.vue'
-import type WinPagination from '@app/components/WinPagination.vue'
+import type WinList from '@app/components/basic/WinList.vue'
+import type WinPagination from '@app/components/basic/WinPagination.vue'
 import type { RatingsResponse } from '@app/types/types'
 import { useI18n } from 'vue-i18n'
